fix number of running processes
scanning /proc/*/stat AND computing other process stats takes too much
time. Consequently, the number of running processes based on the occurences
of 'R' as character indicating the running state is typically zero. Due
to processes are actually changing state during the evaluation of it's
stat(s).

The 'procs_running' number in /proc/stat on the other hand is more
accurate, and can be retrieved in one single 'read' call.

In the unlikely case that procs_running can not be retrieved from
/proc/stat, (read interrupted prior to reading up to 'procs_running'
the value '1' assumed, since this plugin is at least being executed.